requirements.txt
streamlit==1.32.0
pdfplumber==0.10.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
chromadb==0.4.22
langchain-core>=0.2.26
langchain-groq==0.1.9